add dynamic-graph-py.hh
