+++ src/factory-py.cc
@@ -4,6 +4,8 @@
 #include <iostream>
 #include <dynamic-graph/factory.h>
 
+#include "dynamic-graph-py.hh"
+
 using dynamicgraph::Entity;
 using dynamicgraph::ExceptionAbstract;
