+++ src/signal-base-py.cc
@@ -12,6 +12,7 @@
 #include <dynamic-graph/pool.h>
 #include <dynamic-graph/factory.h>
 
+#include "dynamic-graph-py.hh"
 #include "convert-dg-to-py.hh"
 #include "exception.hh"
 #include "signal-wrapper.hh"
